:truck: rename wide-popup spec and fix new-tab navigation test
Button opens about page via window.open (new tab), so test now
captures the new tab via context.waitForEvent instead of checking
the current page URL.
